Conflicts:
	WIP/Sources/FlyAwayPlus/FlyAwayPlus/Controllers/PostController.cs
	WIP/Sources/FlyAwayPlus/FlyAwayPlus/Helpers/GraphDatabaseHelpers.cs
	WIP/Sources/FlyAwayPlus/FlyAwayPlus/Views/LookAround/Index.cshtml

+++ WIP/Sources/FlyAwayPlus/FlyAwayPlus/Controllers/PostController.cs
@@ -106,8 +106,8 @@ namespace FlyAwayPlus.Controllers
         public RedirectToRouteResult Add(FormCollection form)
         {
             string message = Request.Form["message"];
-            Decimal latitude = Decimal.Parse(Request.Form["lat"]);
-            Decimal longitude = Decimal.Parse(Request.Form["lng"]);
+            Double latitude = Double.Parse(Request.Form["lat"]);
+            Double longitude = Double.Parse(Request.Form["lng"]);
             string location = Request.Form["formatted_address"];
             List<string> images = Request.Form["uploadedimages"].Split('#').ToList();
             images.RemoveAt(0);
